# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        description: Details is present only for VALIDATION_ERROR responses.
      message:
        example: Human-readable error description.
        type: string
    type: object
...
  listings.DraftCreateRequest:
    properties:
      type:
        type: string
    required:
    - type
    type: object
  listings.DraftCreateResponse:
    properties:
      current_step:
        type: integer
      draft_id:
        type: string
    type: object
...
  listings.ListingPublicSchema:
    properties:
      allow_children:
        type: boolean
      allow_parties:
        type: boolean
      allow_pets:
        type: boolean
      allow_smoking:
        type: boolean
      amenities:
        items:
          type: string
        type: array
      bathrooms_count:
        type: integer
      beds_count:
        type: integer
      checkin_from:
        type: string
      checkout_until:
        type: string
      created_at:
        type: string
      currency:
        type: string
      deposit_required:
        type: boolean
      description:
        type: string
      floor:
        type: integer
      host:
        $ref: '#/definitions/listings.HostSchema'
      id:
        type: string
      max_guests:
        type: integer
      media:
        items:
          type: string
        type: array
      min_nights:
        type: integer
      name:
        type: string
      price_per_night:
        type: number
      rooms_count:
        type: integer
      square:
        type: number
      total_floors:
        type: integer
      type:
        type: string
      with_invoicing:
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact:
    email: dev@flickey.com
    name: Flickey Dev Team
  description: Property-rental platform backend. Provides OTP-based authentication,
    multi-step listing wizard, media upload workflow, amenity catalogue, and an admin
    moderation panel.
  license:
    name: MIT
  termsOfService: http://flickey.com/terms/
  title: Flickey Go API
  version: "1.0"
paths:
...
  /admin/listings:
    get:
      description: Returns a list of listings, optionally filtered by status.
      parameters:
      - description: Listing status filter (pending, active, rejected, suspended,
          archived)
        enum:
        - pending
        - active
        - rejected
        - suspended
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of listings with host details
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all listings
      tags:
      - Admin
  /admin/listings/{id}:
    get:
      description: Returns full listing details for moderation.
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: LISTING_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get listing details
      tags:
      - Admin
...
  /listings:
    get:
      description: Returns all currently published listings. No authentication required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/listings.ListingPublicSchema'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Browse published listings
      tags:
      - Listings
  /listings/{listing_id}:
    get:
      description: Returns the public-facing details of a specific listing by UUID.
        No authentication required.
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: listing_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listings.ListingPublicSchema'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: LISTING_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Get a single published listing
      tags:
      - Listings
  /listings/drafts:
    post:
      consumes:
      - application/json
      description: Initialises a new multi-step wizard draft with the property type.
        Returns the draft ID and current step number.
      parameters:
      - description: Property type (apartment | house | manor)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/listings.DraftCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/listings.DraftCreateResponse'
        "400":
          description: VALIDATION_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: PROFILE_INCOMPLETE
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new listing draft (step 1)
      tags:
      - Listings
  /listings/drafts/{draft_id}:
    get:
      description: Fetches the full state of a listing draft. Only the owning host
        can access it.
      parameters:
      - description: Draft UUID
        format: uuid
        in: path
        name: draft_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listings.DraftDetailResponse'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: PROFILE_INCOMPLETE
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: DRAFT_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a draft
      tags:
      - Listings
  /listings/drafts/{draft_id}/step-2:
    patch:
      consumes:
      - application/json
      description: 'Sets name, area, floor, guest count, rooms, beds, and bathrooms.
        All fields required. name: 10–100 chars; square: 10–1000 m²; floor/total_floors:
        1–150; max_guests: 1–50; rooms/beds: 1–30; bathrooms: 1–20.'
      parameters:
      - description: Draft UUID
        format: uuid
        in: path
        name: draft_id
        required: true
        type: string
      - description: Property details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/listings.DraftStep2Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listings.DraftStepResponse'
        "400":
          description: VALIDATION_ERROR / INVALID_STEP_TRANSITION
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: PROFILE_INCOMPLETE
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: DRAFT_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update draft – step 2 (property details)
      tags:
      - Listings
  /listings/drafts/{draft_id}/step-3:
    patch:
      consumes:
      - application/json
      description: Attaches 5–15 uploaded media UUIDs to the draft. All referenced
        media must belong to the current user and be in uploaded status.
      parameters:
      - description: Draft UUID
        format: uuid
        in: path
        name: draft_id
        required: true
        type: string
      - description: Array of media UUIDs (5–15 items)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/listings.DraftStep3Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listings.DraftStepResponse'
        "400":
          description: VALIDATION_ERROR / INVALID_STEP_TRANSITION
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: PROFILE_INCOMPLETE / MEDIA_NOT_OWNED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: DRAFT_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update draft – step 3 (media)
      tags:
      - Listings
...
  /listings/my:
    get:
      description: Returns all published (and pending) listings owned by the authenticated
        host.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/listings.ListingHostReadSchema'
            type: array
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: PROFILE_INCOMPLETE
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get host's own listings
      tags:
      - Listings
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: **Bearer {token}**'
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: HttpOnly refresh token cookie set by /auth/verify-otp and rotated
      by /auth/refresh.
    in: cookie
    name: refresh_token
    type: apiKey
swagger: "2.0"